requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==4.2
django-allauth==0.61.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.6
importlib_metadata==7.1.0
Markdown==3.6
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
# psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
pypng==0.20220715.0
python3-openid==3.2.0
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==2.0.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
zipp==3.18.1
